fix(sac): move METRIC_TYPES out of "use server" file
Next.js 14 enforces that "use server" files only export async functions.
METRIC_TYPES is a plain const array (not a function), which causes a
runtime error: 'A "use server" file can only export async functions, found object'.

Changes:
- Create alert-actions.types.ts with METRIC_TYPES (no "use server" directive)
- Remove METRIC_TYPES from alert-actions.ts ("use server" file)
- Update alerts-panel.tsx to import METRIC_TYPES from the new types file

Root cause: PR #402 (Logging Overhaul) added "use server" to alert-actions.ts
without moving the non-function exports out first.
